+++ scvi/data/fields/_protein.py
@@ -7,8 +7,8 @@ from anndata import AnnData
 from mudata import MuData
 
 from ._layer_field import LayerField
+from ._matrix_field import ObsmField
 from ._mudata import BaseMuDataWrapperClass, MuDataWrapper
-from ._obsm_field import ObsmField
 
 logger = logging.getLogger(__name__)
